Separate hasFlag into its own file
